create isomorphic journeys file
