+++ packages/backend-lib/src/journeys.ts
@@ -5,13 +5,9 @@ import { schemaValidate } from "isomorphic-lib/src/resultHandling/schemaValidati
 import { err, ok, Result } from "neverthrow";
 
 import prisma from "./prisma";
-import {
-  EnrichedJourney,
-  JourneyBodyNode,
-  JourneyDefinition,
-  JourneyNodeType,
-  JourneyResource,
-} from "./types";
+import { EnrichedJourney, JourneyDefinition, JourneyResource } from "./types";
+
+export * from "isomorphic-lib/src/journeys";
 
 export function enrichJourney(
   journey: Journey
@@ -27,36 +23,6 @@ export function enrichJourney(
     ...journey,
     definition: definitionResult.value,
   });
-}
-
-function nodeToSegment(node: JourneyBodyNode): string | null {
-  switch (node.type) {
-    case JourneyNodeType.SegmentSplitNode: {
-      return node.variant.segment;
-    }
-    case JourneyNodeType.ExperimentSplitNode:
-      return null;
-    case JourneyNodeType.RateLimitNode:
-      return null;
-    case JourneyNodeType.MessageNode:
-      return null;
-    case JourneyNodeType.DelayNode:
-      return null;
-  }
-}
-
-export function getSubscribedSegments(
-  definition: JourneyDefinition
-): Set<string> {
-  const subscribedSegments = new Set<string>();
-  subscribedSegments.add(definition.entryNode.segment);
-  for (const node of definition.nodes) {
-    const segment = nodeToSegment(node);
-    if (segment) {
-      subscribedSegments.add(segment);
-    }
-  }
-  return subscribedSegments;
 }
 
 type FindManyParams = Parameters<PrismaClient["journey"]["findMany"]>[0];
